+++ tests/sql/schema.sql
@@ -1,3 +1,0 @@
-CREATE EXTENSION postgis;
-
-CREATE EXTENSION postgis_topology;